Expand nodes on clicked location, avoid duplication
Previously, if you were to expand a node, what it would do is add all of the nodes in the top left (not really ideal). Instead, we can take the rendered position of the mouse and expand the nodes using that :).

In addition, by adding a property of "expanded" to each node, we can check if we've already expanded it to avoid duplicated expansion of scalars (which was UGLY!).